refactor : ChampionshipForm 불필요한 함수들 통합했습니다
