+++ src/1_Page/ChampionshipEdit/model/useChampionshipForm.ts
@@ -3,8 +3,15 @@ import usePostChampionship from "../../../3_Entity/Community/usePostChampionship
 import usePutChampionship from "../../../3_Entity/Community/usePutChampionship";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { defaultValues, schema } from "../lib/schema";
-import { convertToAPIChampionship } from "../lib/convert";
+import {
+  convertToAPIChampionship,
+  convertToChampionshipForm,
+} from "../lib/convert";
 import { errorTabDetector } from "../lib/errors";
+import { useSearchParams } from "react-router-dom";
+import React from "react";
+import useGetChampionshipEndData from "../../../3_Entity/Championship/useGetChampionshipEndData";
+import useGetChampionshipInfo from "../../../3_Entity/Championship/useGetChampionshipInfo";
 
 const useChanpionshipForm = (props: UseChanpionshipFormProps) => {
   const { isEditMode, communityIdx, setActiveTab } = props;
@@ -12,9 +19,29 @@ const useChanpionshipForm = (props: UseChanpionshipFormProps) => {
     resolver: yupResolver(schema),
     defaultValues,
   });
+  const { reset } = method;
 
   const [postChampionship] = usePostChampionship(communityIdx);
   const [putChampionship] = usePutChampionship(communityIdx);
+
+  // edit 모드인 경우 championshipIdx는 0으로 설정해 데이터 요청하지 않는다
+  const [searchParams] = useSearchParams();
+  const championshipIdx = isEditMode
+    ? 0
+    : Number(searchParams.get("championshipIdx") ?? 0);
+  const [championshipEndData] = useGetChampionshipEndData(championshipIdx);
+  const [championshipInfo] = useGetChampionshipInfo(championshipIdx);
+  React.useEffect(() => {
+    if (
+      !championshipInfo ||
+      Object.keys(championshipInfo).length === 0 ||
+      !championshipEndData ||
+      Object.keys(championshipEndData).length === 0
+    )
+      return;
+    reset(convertToChampionshipForm(championshipInfo, championshipEndData));
+  }, [championshipEndData, championshipInfo, reset]);
+
   // 성공 시 처리
   const onValid: SubmitHandler<ChampionshipFormValues> = (data) => {
     const body = convertToAPIChampionship(data);
