--- src/1_Page/ChampionshipEdit/ui/EditRequest/type.d.ts
@@ -1,3 +1,0 @@
-type EditRequestProps = {
-  reset: UseFormReset<ChampionshipFormValues>;
-};
